// src/interfaces/OrderTransaction.ts
export interface iBodyOrderTransactionCreate {
  accountCode: number
  salesmanCode: number
  plant: string
  storageLocation: string
  country: string
}